net: make bind_socket() more generic to accomodate client
The client does not invoke the bind() and setsockopt() system calls. It
also requires that the valid addrinfo struct from the list be maintained
for further use. Therefore the bind_socket() function, which was earlier
tailored to the server's needs, is made more generic to accomodate both
the server and the client.