message: write trim function and simply (un-)pack client functions
The trim function cuts off the string after encountering the next-line
character, replacing it by the null terminator. By doing so, it ensures
that all strings are how strings are supposed to be.

The functions to pack and unpack client messages now assume that the
string is already trimmed, therefore those functions are simplified.
Functions dealing with user messages already made that assumption when
written.