message: purge useless parameters and simplify functions
It is hereby assumed that there are no problems related to sizing, due
to care having been taken before getting to the stage of packing and
unpacking messages. Therefore it is pointless to have any of the
arguments relating to sizing, and a lot of if-statements are redundant
as well. All of them have been removed.