server: handle message based on type in a separate function
The code in the main() function to handle messages has been moved to a
separate handle_msg() function, which in turn calls handle_usr_msg() or
handle_client_msg() based on the type determined.

handle_client_msg() determines the subject of the client's request and
executes the response, while handle_usr_msg() (for now) just prepares
the string to print to the terminal.