server: client: enable receiving of user messages in the same thread
So far, messages were only sent to the server but not received.
This commit implements the sending of user messages by the server to all
other clients after the receipt from some client. This occurs in the
same thread as the sending operation.

Messages received by the client are not yet printed to the screen.

Note that the server sending the message to itself causes it to exit, so
never send the message to pfds[0].fd.