client: make "name" statically allocated global variable
The pthread_create() function accepts only one argument for the thread
function. There are 2 arguments from main() that send() and recv()
require: name and sockfd. Instead of creating a new pthread_args struct,
it is easier to just make name a global variable and pass sockfd as the
only argument.