ui: implement function to get the input from ip_win
There are 2 places where input is fetched from ip_win:
	1. login_form()
	2. thread_send()
In both, the operation is the same: check for backspaces and stuff. If
it is a regular, print-able character that is entered, print it and add
it to the buffer.

So avoid code duplication by moving the whole operation into a separate
function and just call that.